chore: do not ignore examples & some fixture folders for renovate
Renovate's `config:recommended` ignores `**/examples/**` by default.
But we want to keep the examples up-to-date, so we need to unignore it.

The same goes for `test/fixtures/file-based-routing`.

I didn't unignore the whole `test/fixtures` folder because the other
fixture, `with-older-espree`, intentionally installs older versions of
dependencies for testing purposes.